Only write files on change.
When formatting files, genemichaels always writes the formatted file to
disk which modifies the file timestamp. This is problematic for tools
that detect whether a file has been reformatted or not.

This commit changes the behavior by comparing the source against the
formatted output and only writing to disk if they are different.

See: https://github.com/andrewbaxter/genemichaels/issues/95